requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy
alembic
passlib
bcrypt
pyjwt
python-jose
python-dotenv